refactor: move global shortcuts
